requirements.txt
determined==0.14.5
determined-cli==0.14.5
determined-deploy==0.14.5
torch==1.8.0
torchvision==0.9.0
jupyterlab==3.0.4
ipywidgets==7.6.3
tensorboard==2.4.0
scipy==1.6.0
tqdm==4.56.0
opencv-python==4.5.1.48
seaborn==0.11.1
facenet-pytorch==2.5.1
echoAI==0.1.3
pytorch-lightning==1.2.3
cbmi-utils==0.1.21
requests==2.25.1
python-dotenv==0.16.0